+++ paths/paths.js
@@ -3,14 +3,14 @@ const path = require("path");
 const rootPath = path.join(__dirname, "..");
 
 const publicPath = path.join(rootPath, "public");
-const utilsPath = path.join(rootPath, "utils");
+// const utilsPath = path.join(rootPath, "utils"); // Use in staticPaths
 const eventsPath = path.join(rootPath, "events");
 
 // Subdirectories in views directory
 const viewsPagesPath = path.join(rootPath, "views/pages");
 const viewsPartialsPath = path.join(rootPath, "views/partials");
 
-const staticPaths = [publicPath, utilsPath, eventsPath];
+const staticPaths = [publicPath, eventsPath];
 const viewsPaths = [viewsPagesPath, viewsPartialsPath];
 
 module.exports = {