+++ utils/generateStoragePath.js
@@ -1,0 +1,6 @@
+const generateStoragePath = (userID, filename, folderID) => {
+  const storagePath = `/${userID}${folderID ? `/${folderID}` : ""}/${filename}`;
+  return storagePath;
+};
+
+module.exports = generateStoragePath;