+++ utils/utils.js
@@ -1,0 +1,13 @@
+const generateRandomIndex = require("./generateRandomIndex");
+const generateStoragePath = require("./generateStoragePath");
+const getFileExtension = require("./getFileExtension");
+const getFilesFromPath = require("./getFilesFromPath");
+const logger = require("./logger");
+
+module.exports = {
+  generateRandomIndex,
+  generateStoragePath,
+  getFileExtension,
+  getFilesFromPath,
+  logger,
+};